fix: create file in folder 'helpers' with functions for Image-resizer
