+++ src/components/event/ImageResizer/Image-resizer.js
@@ -6,6 +6,12 @@ import PropTypes from "prop-types";
 import "./image-resizer.css";
 import Button from "@material-ui/core/Button";
 import imageConstants from "../../../constants/ImageResizer";
+import {
+  cropImage,
+  onCropChange,
+  onZoomChange,
+  onCropComplete,
+} from "../../helpers/Image-resizer";
 
 const { CROP, ROUND } = imageConstants;
 
@@ -21,7 +27,6 @@ const ImageResizer = ({
   const [aspect, setAspect] = useState(null);
   const [showGrid, setShowGrid] = useState(null);
   const [croppedArea, setCroppedArea] = useState(null);
-
   useEffect(() => {
     const isRound = cropShape === ROUND;
     setAspect(isRound ? 1 : 16 / 9);
@@ -30,58 +35,6 @@ const ImageResizer = ({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  const onCropChange = crop => setCrop(crop);
-
-  const onCropComplete = (croppedArea, croppedAreaPixels) =>
-    setCroppedArea(croppedAreaPixels);
-
-  const onZoomChange = zoom => setZoom(zoom);
-
-  const createImage = url =>
-    new Promise((resolve, reject) => {
-      const image = new Image();
-      image.addEventListener("load", () => resolve(image));
-      image.addEventListener("error", error => reject(error));
-      image.src = url;
-    });
-
-  const getCroppedImg = async (imageSrc, pixelCrop) => {
-    const image = await createImage(imageSrc);
-    const canvas = document.createElement("canvas");
-    const ctx = canvas.getContext("2d");
-
-    const maxSize = Math.max(image.width, image.height);
-    const safeArea = 2 * ((maxSize / 2) * Math.sqrt(2));
-
-    canvas.width = safeArea;
-    canvas.height = safeArea;
-
-    ctx.translate(safeArea / 2, safeArea / 2);
-    ctx.translate(-safeArea / 2, -safeArea / 2);
-
-    ctx.drawImage(
-      image,
-      safeArea / 2 - image.width * 0.5,
-      safeArea / 2 - image.height * 0.5,
-    );
-    const data = ctx.getImageData(0, 0, safeArea, safeArea);
-
-    canvas.width = pixelCrop.width;
-    canvas.height = pixelCrop.height;
-
-    ctx.putImageData(
-      data,
-      Math.round(0 - safeArea / 2 + image.width * 0.5 - pixelCrop.x),
-      Math.round(0 - safeArea / 2 + image.height * 0.5 - pixelCrop.y),
-    );
-
-    return canvas.toDataURL("image/jpeg");
-  };
-
-  const cropImage = async () => {
-    const croppedImage = await getCroppedImg(image.preview, croppedArea);
-    handleOnCrop([croppedImage], onChange);
-  };
   return (
     <div className="ImageResizer">
       <div className="crop-container">
@@ -90,9 +43,11 @@ const ImageResizer = ({
           crop={crop}
           zoom={zoom}
           aspect={aspect}
-          onCropChange={onCropChange}
-          onCropComplete={onCropComplete}
-          onZoomChange={onZoomChange}
+          onCropChange={crop => onCropChange(crop, setCrop)}
+          onCropComplete={(croppedArea, croppedAreaPixels) =>
+            onCropComplete(croppedArea, croppedAreaPixels, setCroppedArea)
+          }
+          onZoomChange={zoom => onZoomChange(zoom, setZoom)}
           cropShape={cropShape}
           showGrid={showGrid}
         />
@@ -104,13 +59,15 @@ const ImageResizer = ({
           max={3}
           step={0.1}
           aria-labelledby="Zoom"
-          onChange={(e, zoom) => onZoomChange(zoom)}
+          onChange={(e, zoom) => onZoomChange(zoom, setZoom)}
         />
         <Button
           type="button"
           color="primary"
           disabled={submitting}
-          onClick={cropImage}
+          onClick={() => {
+            cropImage(image, croppedArea, handleOnCrop, onChange);
+          }}
           style={{ float: "right" }}
         >
           {CROP}
