+++ src/components/helpers/Image-resizer.js
@@ -1,0 +1,55 @@
+export const onCropChange = (crop, setCrop) => setCrop(crop);
+
+export const onZoomChange = (zoom, setZoom) => setZoom(zoom);
+
+export const onCropComplete = (
+  croppedArea,
+  croppedAreaPixels,
+  setCroppedArea,
+) => setCroppedArea(croppedAreaPixels);
+
+const createImage = url =>
+  new Promise((resolve, reject) => {
+    const image = new Image();
+    image.addEventListener("load", () => resolve(image));
+    image.addEventListener("error", error => reject(error));
+    image.src = url;
+  });
+
+const getCroppedImg = async (imageSrc, pixelCrop) => {
+  const image = await createImage(imageSrc);
+  const canvas = document.createElement("canvas");
+  const ctx = canvas.getContext("2d");
+
+  const maxSize = Math.max(image.width, image.height);
+  const safeArea = 2 * ((maxSize / 2) * Math.sqrt(2));
+
+  canvas.width = safeArea;
+  canvas.height = safeArea;
+
+  ctx.translate(safeArea / 2, safeArea / 2);
+  ctx.translate(-safeArea / 2, -safeArea / 2);
+
+  ctx.drawImage(
+    image,
+    safeArea / 2 - image.width * 0.5,
+    safeArea / 2 - image.height * 0.5,
+  );
+  const data = ctx.getImageData(0, 0, safeArea, safeArea);
+
+  canvas.width = pixelCrop.width;
+  canvas.height = pixelCrop.height;
+
+  ctx.putImageData(
+    data,
+    Math.round(0 - safeArea / 2 + image.width * 0.5 - pixelCrop.x),
+    Math.round(0 - safeArea / 2 + image.height * 0.5 - pixelCrop.y),
+  );
+
+  return canvas.toDataURL("image/jpeg");
+};
+
+export const cropImage = async (image, croppedArea, handleOnCrop, onChange) => {
+  const croppedImage = await getCroppedImg(image.preview, croppedArea);
+  handleOnCrop([croppedImage], onChange);
+};
